backend/requirements.txt
# Core Dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ML/CV
torch==2.2.0
torchvision==0.17.0
ultralytics==8.1.0
timm==0.9.12
transformers==4.37.0
opencv-python==4.9.0.80
pillow==10.2.0
albumentations==1.3.1
scikit-learn==1.4.0

# Vector Search
faiss-cpu==1.7.4
numpy==1.26.3

# Utils
tqdm==4.66.1
